+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Added
 - `Global_exception_handler` to `__all__` in `src/packer/__init__.py` to expose the exception handler module. 
 - `global_exception_handler` function to a `Global_exception_handler` class to be able to set it up with the necessary data and run it with it's predefined actions. 
+- Warning filter to ignore AVX2 capability warnings from pygame. 
 
 ### Changed
 - All use cases and modules to import from config.py the handler which is imported from exceptions.py and just set up and configured in config.py.